+++ sketch-extensions/src/androidTest/java/com/github/panpf/sketch/extensions/test/decode/ApkIconBitmapDecoderTest.kt
@@ -46,7 +46,7 @@ import org.junit.runner.RunWith
 class ApkIconBitmapDecoderTest {
 
     @Test
-    fun testSupportAppIcon() {
+    fun testSupportApkIcon() {
         ComponentRegistry.Builder().apply {
             build().apply {
                 Assert.assertEquals(